Make aggregate web configuration conditional
 - Will be configured only when web enabled in classpath
