+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/aggregate/AggregateApplicationBuilder.java
@@ -16,42 +16,24 @@
 
 package org.springframework.cloud.stream.aggregate;
 
-import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.regex.Pattern;
 
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.BeanFactoryUtils;
-import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 import org.springframework.beans.factory.SmartInitializingSingleton;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.actuate.autoconfigure.EndpointCorsProperties;
-import org.springframework.boot.actuate.autoconfigure.ManagementServerProperties;
 import org.springframework.boot.actuate.autoconfigure.ManagementServerPropertiesAutoConfiguration;
-import org.springframework.boot.actuate.endpoint.AbstractEndpoint;
 import org.springframework.boot.actuate.endpoint.MetricReaderPublicMetrics;
 import org.springframework.boot.actuate.endpoint.MetricsEndpoint;
-import org.springframework.boot.actuate.endpoint.mvc.AbstractEndpointMvcAdapter;
-import org.springframework.boot.actuate.endpoint.mvc.AbstractMvcEndpoint;
-import org.springframework.boot.actuate.endpoint.mvc.EndpointHandlerMapping;
-import org.springframework.boot.actuate.endpoint.mvc.EndpointHandlerMappingCustomizer;
-import org.springframework.boot.actuate.endpoint.mvc.MvcEndpoint;
-import org.springframework.boot.actuate.endpoint.mvc.MvcEndpointSecurityInterceptor;
-import org.springframework.boot.actuate.endpoint.mvc.MvcEndpoints;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
-import org.springframework.boot.autoconfigure.web.BasicErrorController;
 import org.springframework.boot.autoconfigure.web.DispatcherServletAutoConfiguration;
 import org.springframework.boot.autoconfigure.web.EmbeddedServletContainerAutoConfiguration;
 import org.springframework.boot.autoconfigure.web.HttpMessageConvertersAutoConfiguration;
@@ -68,16 +50,11 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Lazy;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.env.PropertySources;
 import org.springframework.integration.monitor.IntegrationMBeanExporter;
 import org.springframework.util.Assert;
-import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.method.HandlerMethod;
-import org.springframework.web.servlet.handler.AbstractHandlerMethodMapping;
-import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
 
 /**
  * Application builder for {@link AggregateApplication}.
@@ -230,9 +207,11 @@ public class AggregateApplicationBuilder implements AggregateApplication, Applic
 		}
 		if (this.parentContext == null) {
 			if (Boolean.TRUE.equals(this.webEnvironment)) {
-				this.addParentSources(new Object[]{EmbeddedServletContainerAutoConfiguration.class, WebMvcAutoConfiguration.class,
-						DispatcherServletAutoConfiguration.class, HttpMessageConvertersAutoConfiguration.class,
-						ManagementServerPropertiesAutoConfiguration.class, ServerPropertiesAutoConfiguration.class});
+				this.addParentSources(
+						new Object[] { EmbeddedServletContainerAutoConfiguration.class, WebMvcAutoConfiguration.class,
+								DispatcherServletAutoConfiguration.class, HttpMessageConvertersAutoConfiguration.class,
+								ManagementServerPropertiesAutoConfiguration.class,
+								ServerPropertiesAutoConfiguration.class });
 			}
 			this.parentContext = AggregateApplicationUtils.createParentContext(
 					this.parentSources.toArray(new Object[0]),
@@ -293,71 +272,22 @@ public class AggregateApplicationBuilder implements AggregateApplication, Applic
 				appConfigurer.args(argsToUpdate.toArray(new String[0]));
 			}
 		}
-		exposeWebEndpoints(apps);
+		try {
+			AggregateWebConfiguration.WebEndpointConfigurer webEndpointConfigurer = this.parentContext
+					.getBean(AggregateWebConfiguration.WebEndpointConfigurer.class);
+			webEndpointConfigurer.exposeWebEndpoints(apps);
+		}
+		catch (BeansException e) {
+			for (int i = apps.size() - 1; i >= 0; i--) {
+				AggregateApplicationBuilder.AppConfigurer<?> appConfigurer = apps.get(i);
+				appConfigurer.embed();
+			}
+		}
 		if (BeanFactoryUtils.beansOfTypeIncludingAncestors(this.parentContext, AggregateApplication.class)
 				.size() == 0) {
 			this.parentContext.getBeanFactory().registerSingleton("aggregateApplicationAccessor", this);
 		}
 		return this.parentContext;
-	}
-
-	private void exposeWebEndpoints(List<AppConfigurer<?>> apps) {
-		Set<MvcEndpoint> mvcEndpointsToAdd = new HashSet<>();
-		Map<RequestMappingInfo, Map<Object, Method>> requestMappingInfo = new HashMap<>();
-		for (int i = apps.size() - 1; i >= 0; i--) {
-			AppConfigurer<?> appConfigurer = apps.get(i);
-			ConfigurableApplicationContext childContext = appConfigurer.embed();
-			// Map the Spring MVC RequestMapping into parent context
-			AbstractHandlerMethodMapping handlerMethodMapping = childContext.getBean(AbstractHandlerMethodMapping.class);
-			Map<RequestMappingInfo, HandlerMethod> mappings = handlerMethodMapping.getHandlerMethods();
-			for (Map.Entry<RequestMappingInfo, HandlerMethod> mapping : mappings.entrySet()) {
-				Map<Object, Method> handlerMethodMap = new HashMap<>();
-				HandlerMethod handlerMethod = mapping.getValue();
-				if (!BasicErrorController.class.isAssignableFrom(handlerMethod.getBeanType())) {
-					Object handlerBean = handlerMethod.getBean();
-					try {
-						this.parentContext.getBeanFactory().getBean(handlerBean.toString());
-					}
-					catch (NoSuchBeanDefinitionException e) {
-						// ChildContext is expected to have the bean as the request mapping is retrieved from the child context.
-						this.parentContext.getBeanFactory().registerSingleton(handlerBean.toString(), childContext.getBean(handlerBean.toString()));
-					}
-					handlerMethodMap.put(handlerBean, handlerMethod.getMethod());
-				}
-				requestMappingInfo.put(mapping.getKey(), handlerMethodMap);
-			}
-			// Map the actuator web endpoints
-			MvcEndpoints mvcEndpoints = new MvcEndpoints();
-			mvcEndpoints.setApplicationContext(childContext);
-			try {
-				mvcEndpoints.afterPropertiesSet();
-			}
-			catch (Exception e) {
-				throw new RuntimeException(e);
-			}
-			for (MvcEndpoint mvcEndpoint : mvcEndpoints.getEndpoints()) {
-				String namespaceValue = Pattern.compile("\\w+").matcher(appConfigurer.getNamespace()).matches() ? appConfigurer.getNamespace() : "child_" + i;
-				if (mvcEndpoint instanceof AbstractMvcEndpoint) {
-					((AbstractMvcEndpoint) mvcEndpoint).setPath("/" + namespaceValue + "/" + mvcEndpoint.getPath());
-				}
-				else if (mvcEndpoint instanceof AbstractEndpointMvcAdapter) {
-					((AbstractEndpointMvcAdapter) mvcEndpoint).setPath("/" + namespaceValue + "/" + mvcEndpoint.getPath());
-				}
-				else if (mvcEndpoint instanceof AbstractEndpoint) {
-					((AbstractEndpoint) mvcEndpoint).setId(namespaceValue + "_" + ((AbstractEndpoint) mvcEndpoint).getId());
-				}
-			}
-			mvcEndpointsToAdd.addAll(mvcEndpoints.getEndpoints());
-		}
-		MvcEndpointsHolder mvcEndpointsHolder = new MvcEndpointsHolder(mvcEndpointsToAdd);
-		this.parentContext.getBeanFactory().registerSingleton("mvcEndpointsHolder", mvcEndpointsHolder);
-		// invoke lazy bean to get instantiated after setting mvcEndpoints
-		EndpointHandlerMapping endpointHandlerMapping = this.parentContext.getBean(EndpointHandlerMapping.class);
-		for (Map.Entry<RequestMappingInfo, Map<Object, Method>> mappingInfoMapEntry : requestMappingInfo.entrySet()) {
-			for (Map.Entry<Object, Method> handlerMethodEntry : mappingInfoMapEntry.getValue().entrySet()) {
-				endpointHandlerMapping.registerMapping(mappingInfoMapEntry.getKey(), handlerMethodEntry.getKey(), handlerMethodEntry.getValue());
-			}
-		}
 	}
 
 	private boolean selfContained() {
@@ -571,17 +501,9 @@ public class AggregateApplicationBuilder implements AggregateApplication, Applic
 	}
 
 	@EnableBinding
-	@ImportAutoConfiguration({ChannelBindingAutoConfiguration.class})
+	@ImportAutoConfiguration({ ChannelBindingAutoConfiguration.class })
+	@Import(AggregateWebConfiguration.class)
 	public static class ParentConfiguration {
-
-		@Autowired(required = false)
-		private ManagementServerProperties managementServerProperties;
-
-		@Autowired(required = false)
-		private List<EndpointHandlerMappingCustomizer> mappingCustomizers;
-
-		@Autowired(required = false)
-		private EndpointCorsProperties corsProperties;
 
 		@Bean
 		@ConditionalOnMissingBean(SharedBindingTargetRegistry.class)
@@ -593,93 +515,6 @@ public class AggregateApplicationBuilder implements AggregateApplication, Applic
 		@ConditionalOnMissingBean(SharedChannelRegistry.class)
 		public SharedChannelRegistry sharedChannelRegistry(SharedBindingTargetRegistry sharedBindingTargetRegistry) {
 			return new SharedChannelRegistry(sharedBindingTargetRegistry);
-		}
-
-		@Bean
-		@Lazy
-		@ConditionalOnClass(name = "org.springframework.boot.actuate.endpoint.mvc.MvcEndpointSecurityInterceptor")
-		public EndpointHandlerMapping endpointHandlerMapping(MvcEndpointsHolder mvcEndpointsHolder) {
-			EndpointHandlerMapping mapping = null;
-			if (this.corsProperties != null) {
-				CorsConfiguration corsConfiguration = getCorsConfiguration(this.corsProperties);
-				mapping = new EndpointHandlerMapping(mvcEndpointsHolder.getMvcEndpoints(), corsConfiguration);
-			}
-			else {
-				mapping = new EndpointHandlerMapping(mvcEndpointsHolder.getMvcEndpoints(), null);
-			}
-			if (managementServerProperties != null) {
-				mapping.setPrefix(this.managementServerProperties.getContextPath());
-				MvcEndpointSecurityInterceptor securityInterceptor = new MvcEndpointSecurityInterceptor(
-						this.managementServerProperties.getSecurity().isEnabled(),
-						this.managementServerProperties.getSecurity().getRoles());
-				mapping.setSecurityInterceptor(securityInterceptor);
-			}
-			if (this.mappingCustomizers != null) {
-				for (EndpointHandlerMappingCustomizer customizer : this.mappingCustomizers) {
-					customizer.customize(mapping);
-				}
-			}
-			return mapping;
-		}
-
-		@Bean
-		@Lazy
-		@ConditionalOnMissingClass("org.springframework.boot.actuate.endpoint.mvc.MvcEndpointSecurityInterceptor")
-		public EndpointHandlerMapping endpointHandlerMappingWithoutSecurityInterceptor(MvcEndpointsHolder mvcEndpointsHolder) {
-			EndpointHandlerMapping mapping = null;
-			if (this.corsProperties != null) {
-				CorsConfiguration corsConfiguration = getCorsConfiguration(this.corsProperties);
-				mapping = new EndpointHandlerMapping(mvcEndpointsHolder.getMvcEndpoints(), corsConfiguration);
-			}
-			else {
-				mapping = new EndpointHandlerMapping(mvcEndpointsHolder.getMvcEndpoints(), null);
-			}
-			if (managementServerProperties != null) {
-				mapping.setPrefix(this.managementServerProperties.getContextPath());
-			}
-			if (this.mappingCustomizers != null) {
-				for (EndpointHandlerMappingCustomizer customizer : this.mappingCustomizers) {
-					customizer.customize(mapping);
-				}
-			}
-			return mapping;
-		}
-
-		private CorsConfiguration getCorsConfiguration(EndpointCorsProperties properties) {
-			if (CollectionUtils.isEmpty(properties.getAllowedOrigins())) {
-				return null;
-			}
-			CorsConfiguration configuration = new CorsConfiguration();
-			configuration.setAllowedOrigins(properties.getAllowedOrigins());
-			if (!CollectionUtils.isEmpty(properties.getAllowedHeaders())) {
-				configuration.setAllowedHeaders(properties.getAllowedHeaders());
-			}
-			if (!CollectionUtils.isEmpty(properties.getAllowedMethods())) {
-				configuration.setAllowedMethods(properties.getAllowedMethods());
-			}
-			if (!CollectionUtils.isEmpty(properties.getExposedHeaders())) {
-				configuration.setExposedHeaders(properties.getExposedHeaders());
-			}
-			if (properties.getMaxAge() != null) {
-				configuration.setMaxAge(properties.getMaxAge());
-			}
-			if (properties.getAllowCredentials() != null) {
-				configuration.setAllowCredentials(properties.getAllowCredentials());
-			}
-			return configuration;
-		}
-	}
-
-	public static class MvcEndpointsHolder {
-
-		private final Set<MvcEndpoint> mvcEndpoints;
-
-		public MvcEndpointsHolder(Set<MvcEndpoint> mvcEndpoints) {
-			this.mvcEndpoints = mvcEndpoints;
-		}
-
-		public Set<MvcEndpoint> getMvcEndpoints() {
-			return mvcEndpoints;
 		}
 	}
 }
